requirements.txt
openai>=1.0.0
aiohttp>=3.8.5
aiofiles>=23.2.1
python-dotenv>=1.0.0
tenacity>=8.2.2
numpy>=1.24.3
sentence-transformers>=2.2.2
PyPDF2>=3.0.1
